Updating int/double
Sensor position comes back as a int, but we are relying on the compiler to cast it as a double correctly.....Not that greatest....and then we are doing division where we are doing int/double....Also relying on the compiler. So I tried to clean all it up. Now setpoints are casted all as int. Note though that the set function requires a double input....So I used static_cast<double> to fix that. Moral of the story: Don't rely on compilers to cast things for you....They are using rules someone else created that you don't know.